Harden input/audio init across all backends
Introduces a broad compatibility and robustness pass for client input and audio: adds ordered input reset events (full and mouse-only), per-binding generation tags for deferred +/- commands, safer voice ownership handling, and stronger focus/device-loss recovery on SDL, Win32, and X11/Linux paths. Reworks text/clipboard handling to use atomic UTF-8 scalar insertion, improves console/chat rendering and chat field ABI handling, and adds stricter pointer/wheel/capture behavior to prevent stuck keys, phantom motion, and lost releases.

Audio initialization is hardened across OpenAL, SDL, WASAPI/DirectSound, ALSA/OSS, and legacy DMA with shared validation policy helpers, better diagnostics, safer format/buffer checks, and improved fallback behavior. Build wiring, docs, and extensive source/unit tests were updated to cover the new reset contracts, input parsing/projection math, joystick paths, and audio init policy behavior.

+++ code/qcommon/vm.c
@@ -2754,18 +2754,23 @@ static intptr_t VM_CallNativeExports( vm_t *vm, int callnum, const intptr_t *arg
 			}
 			return VM_NormalizeQbooleanResult(
 				((qboolean (QDECL *)( int, void * ))exportFunc)( (int)args[0], (void *)args[1] ) );
+		// Retail cgamex86.dll returns the chat-field origin and width as x87
+		// floats in ST(0), not as integers in EAX.  Calling them through an int
+		// prototype reads a stale EAX and leaves the result stranded on the FPU
+		// stack, which places the chat overlay at an arbitrary offscreen Y and
+		// eventually overflows x87 after eight frames.
 		case CG_GET_CHAT_FIELD_Y:
 			exportFunc = dllExports[CG_NATIVE_EXPORT_GET_CHAT_FIELD_Y];
 			if ( !exportFunc ) {
 				break;
 			}
-			return ((int (QDECL *)( void ))exportFunc)();
+			return (intptr_t)((float (QDECL *)( void ))exportFunc)();
 		case CG_GET_CHAT_FIELD_PIXEL_WIDTH:
 			exportFunc = dllExports[CG_NATIVE_EXPORT_GET_CHAT_FIELD_PIXEL_WIDTH];
 			if ( !exportFunc ) {
 				break;
 			}
-			return ((int (QDECL *)( void ))exportFunc)();
+			return (intptr_t)((float (QDECL *)( void ))exportFunc)();
 		case CG_GET_CHAT_FIELD_WIDTH_IN_CHARS:
 			exportFunc = dllExports[CG_NATIVE_EXPORT_GET_CHAT_FIELD_WIDTH_IN_CHARS];
 			if ( !exportFunc ) {
